modules work saved here
Curently running into a g++ bug when trying to use spdlog in the logger fiel. See following link for potential info:
https://stackoverflow.com/questions/76808137/using-spdlog-with-modules-c20-gives-internal-compiler-error-msvc-2022

I do not like abandoning this, but if the goal is to make a game engine then it is unreasonable to be stuck on this for
so long. This may be something to return to at some point if we want to work on making this functional within the compiler,
but I've heard pushing into g++ is a pain at best.