add window architecture
Window should only display empty window but is failing in unique_ptr ctor right now.
Architecturally, a manager object contains the sole Window object as well as a WindowEventListener
that receives events.

Question: shouldn't the manager be the one to pop events off the listener queue?
otherwise how is the WindowEventListener class supposed to communicate to the manager?
I suppose a basic API is to just fetch the front event and pass it off to the manager (like a
getFirstEvent() function that cna be added to EventListener).